+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* `react-container-query`
+  * `ContainerQuery` now accepts the render function through the `children` prop
+  * `ResizeObserver` does the same
+
 ## [2.4.2]
 
 ### Fixed